requirements.txt
asgiref==3.8.1
Django==4.2.20
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
importlib_metadata==8.6.1
Markdown==3.8
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
zipp==3.21.0
drf-yasg==1.21.7
faker==37.1.0
setuptools==78.1.0
django-filter==25.1
pytest-django
requests